+++ src/ui/components/Page/articles.js
@@ -1,22 +1,48 @@
 import React, { useEffect, useState } from 'react'
 
 import { ipfs } from '../../common/ipfs'
+import { loadHTML } from '../../common/file'
 import { Spinner } from '../Spinner'
 
 import css from './articles.css'
 
 export const Articles = () => {
   const [loading, setLoading] = useState(true)
-  const [pinset, setPinset] = useState([])
+  const [articles, setArticles] = useState([])
 
   useEffect(() => {
     const getPinset = async () => {
       if (ipfs) {
-        const newPinset = await ipfs.pin.ls()
-        setPinset(newPinset)
-        setLoading(false)
+        const pinset = await ipfs.pin.ls()
+        const articles = (await Promise.all(
+          pinset.map(async data => {
+            try {
+              const { hash } = data
 
-        newPinset.map(({ hash }) => {})
+              // Parse entry html
+              const $ = await loadHTML(hash)
+              return $
+                ? {
+                    html: $,
+                    title: $('title').text(),
+                    description: $("meta[name='description']").attr('content'),
+                    author: $("meta[property='article:author']").attr(
+                      'content'
+                    ),
+                    timestamp: $("time[itemprop='datePublished']").attr(
+                      'datetime'
+                    )
+                  }
+                : null
+            } catch (error) {
+              console.error(error)
+              return null
+            }
+          })
+        )).filter(article => article)
+
+        setArticles(articles)
+        setLoading(false)
       }
     }
     getPinset()
@@ -31,7 +57,7 @@ export const Articles = () => {
           text="Articles loading ..."
         />
       )}
-      {!loading && <div> {JSON.stringify(pinset)}</div>}
+      {!loading && <div> {JSON.stringify(articles)}</div>}
     </section>
   )
 }
