+++ src/ui/components/Page/explore.js
@@ -1,7 +1,7 @@
 import * as cheerio from 'cheerio'
 import React, { useContext, useEffect, useState } from 'react'
 
-import { bufferToDataURI } from '../../common/file'
+import { bufferToDataURI, loadHTML } from '../../common/file'
 import { ipfs } from '../../common/ipfs'
 import { PinButton } from '../Button/pin'
 import { HashContext } from '../Context/hash'
@@ -18,45 +18,16 @@ export const Explore = () => {
 
   const [html, setHtml] = useState('')
 
-  const get = hash => {
-    ipfs.get(hash, (err, files) => {
-      if (err) {
-        console.log(err)
-        return undefined
-      }
-
-      if (files && files.length > 0) {
-        // Filter data
-        const data = files.reduce((sum, file) => {
-          const { path, content } = file
-          const cleannedPath = path.replace(`${hash}/`, '')
-          if (cleannedPath && content) {
-            sum[`${cleannedPath}`] = content
-          }
-          return sum
-        }, {})
-
-        // Parse entry html
-        const $ = cheerio.load(data['index.html'].toString('utf8'), {
-          decodeEntities: false,
-          xmlMode: true
-        })
-        $('img').each((index, image) => {
-          const element = $(image)
-          const src = element.attr('src')
-          element.attr('src', bufferToDataURI(data[src]))
-        })
-        setHtml($('body').html())
-      }
-      setLoading(false)
-    })
-  }
-
   useEffect(() => {
     if (hash) {
       setHtml('')
       setLoading(true)
-      get(hash)
+      loadHTML(hash).then($ => {
+        if ($) {
+          setHtml($('body').html())
+          setLoading(false)
+        }
+      })
     }
   }, [hash])
